requirements.txt
aiogram==3.5.0
SQLAlchemy==1.4.48
asyncpg==0.27.0
alembic==1.10.3
redis>=4.5
phonenumbers==8.13.14
python-dotenv==1.0.1
python-multipart==0.0.6
psycopg[binary]==3.1.12
pydantic[email]
